Move tests to tests/ folder and add CI/CD workflow
- Move all test files (publicLog, buildUserMention, isForwardedMessage) to tests/ directory
- Update import paths to reference parent directory modules
- Update package.json test script to target tests/ folder
- Add GitHub Actions workflow for running tests on PRs and default branch
- Tests run on Node.js across ubuntu, macos, and windows

Following best practices from link-foundation/js-ai-driven-development-pipeline-template
