+++ tests/isForwardedMessage.test.js
